fix: 데이터 구조 불일치 해결

+++ src/main/java/org/swbe/domain/lostitem/entity/OfficeStaffAssignment.java
@@ -35,5 +35,5 @@ public class OfficeStaffAssignment {
   private LocalDateTime endedAt;
 
   @Column(name = "active_marker", insertable = false, updatable = false)
-  private Integer activeMarker;
+  private Byte activeMarker;
 }
